requirements.txt
agno==1.1.13
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.0
deepeval==2.4.3
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.16
docx2txt==0.8
dspy==3.0.3
duckduckgo_search==7.5.2
durationpy==0.9
execnet==2.1.1
executing==2.0.1
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
groq==0.19.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
lancedb==0.21.1
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.39
langchain-huggingface==0.1.2
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.10
litellm==1.61.11
llama-cloud==0.1.13
llama-cloud-services==0.6.1
llama-index==0.12.19
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.19
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
milvus-lite==2.4.11
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.30.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.66.3
pandas==2.2.3
peewee==3.17.9
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pluggy==1.5.0
portalocker==3.1.1
posthog==3.16.0
primp==0.14.0
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.0
pydantic==2.10.6
pydantic-settings==2.8.0
pydeck==0.9.1
pylance==0.24.1
pymilvus==2.5.4
pypdf==5.3.0
pytest==8.3.4
pytest-repeat==0.9.3
pytest-xdist==3.6.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentry-sdk==2.22.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
streamlit==1.43.2
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tantivy==0.22.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.6.0
tornado==6.4
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing-inspect==0.9.0
# typing_extensions==4.11.0
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
zstandard==0.23.0
litellm==1.61.11
ddgs>=9.0.0
python-dotenv
dspy